move all custom files into plugin
