--- studio/components/StudioLayout.tsx
@@ -1,10 +1,0 @@
-import {LayoutProps} from 'sanity'
-import {SupportedLanguagesContextProvider} from './SupportedLanguages/hooks/useSupportedLanguages'
-
-export const StudioLayout = (props: LayoutProps) => {
-  return (
-    <SupportedLanguagesContextProvider>
-      {props.renderDefault(props)}
-    </SupportedLanguagesContextProvider>
-  )
-}
