--- studio/plugins/TranslationService/components/TranslationServiceToolComponent.tsx
@@ -1,3 +1,0 @@
-export const TranslationServiceToolComponent = () => {
-  return <div>TranslationServiceToolComponent</div>
-}
