+++ studio/plugins/TranslationService/index.ts
@@ -1,17 +1,31 @@
 import {definePlugin} from 'sanity'
-import {translationServiceTool} from './tool/translationServiceTool'
 import {PluginConfig} from './types'
 import {DEFAULT_CONFIG} from './constants'
 import {TranslationServiceProvider} from './components'
+import {TranslationServiceAction} from './actions'
+import {SanityDocumentInputComponent} from './components/SanityDocumentInputComponent'
+import {schemaTypes} from './schemas'
 
 export const translationService = definePlugin<PluginConfig>((config) => {
   const pluginConfig = {...DEFAULT_CONFIG, ...config}
   return {
     name: 'translationService',
-    tools: (prev) => [...prev, translationServiceTool],
     studio: {
       components: {
         layout: (props) => TranslationServiceProvider({...props, pluginConfig}),
+      },
+    },
+    document: {
+      actions: (prev, context) => {
+        return context.schemaType === 'post' ? [...prev, TranslationServiceAction] : prev
+      },
+    },
+    schema: {
+      types: schemaTypes,
+    },
+    form: {
+      components: {
+        input: SanityDocumentInputComponent,
       },
     },
   }
