+++ studio/plugins/TranslationService/schemas/index.ts
@@ -1,0 +1,5 @@
+import localeBlockContent from './localeBlockContent'
+import localeTitle from './localeTitle'
+import supportedLanguages from './supportedLanguages'
+
+export const schemaTypes = [localeTitle, localeBlockContent, supportedLanguages]
