--- studio/plugins/TranslationService/tool/translationServiceTool.ts
@@ -1,10 +1,0 @@
-import {RocketIcon} from '@sanity/icons'
-import {TranslationServiceToolConfig} from '../types'
-import {TranslationServiceToolComponent} from '../components'
-
-export const translationServiceTool: TranslationServiceToolConfig = {
-  title: 'Translations',
-  name: 'translations',
-  icon: RocketIcon,
-  component: TranslationServiceToolComponent,
-}
