+++ studio/sanity.config.ts
@@ -1,20 +1,14 @@
-import {defineConfig, SchemaPluginOptions, StudioComponentsPluginOptions} from 'sanity'
+import {defineConfig, SchemaPluginOptions} from 'sanity'
 import {deskTool} from 'sanity/desk'
 import {visionTool} from '@sanity/vision'
 import {schemaTypes} from './schemas'
 import {structure} from './structure'
 import {defaultTemplates} from './schemas/config/defaultTemplates'
-import {ManualTranslateAction} from './actions'
-import {StudioLayout} from './components/StudioLayout'
 import {translationService} from './plugins/TranslationService'
 
 const schema: SchemaPluginOptions = {
   types: schemaTypes,
   templates: (prev) => [...defaultTemplates, ...prev],
-}
-
-const studioComponents: StudioComponentsPluginOptions = {
-  layout: StudioLayout,
 }
 
 const baseConfig = {
@@ -29,14 +23,6 @@ const baseConfig = {
     }),
   ],
   schema,
-  document: {
-    actions: (prev, context) => {
-      return context.schemaType === 'post' ? [...prev, ManualTranslateAction] : prev
-    },
-  },
-  studio: {
-    components: studioComponents,
-  },
 }
 
 export default defineConfig([
