+++ studio/schemas/index.ts
@@ -2,16 +2,5 @@ import blockContent from './blockContent'
 import category from './category'
 import post from './post'
 import author from './author'
-import supportedLanguages from './supportedLanguages'
-import localeTitle from './localeTitle'
-import localeBlockContent from './localeBlockContent'
 
-export const schemaTypes = [
-  post,
-  author,
-  category,
-  blockContent,
-  supportedLanguages,
-  localeTitle,
-  localeBlockContent,
-]
+export const schemaTypes = [post, author, category, blockContent]
