+++ studio/schemas/post.ts
@@ -1,8 +1,7 @@
 import {ConditionalPropertyCallbackContext, defineField, defineType} from 'sanity'
 import {ComposeIcon} from '@sanity/icons'
 import {uniqueSlugByLanguage} from '../utils/uniqueSlugByLanguage'
-import {PostDocumentInput} from '../components/PostDocumentInput'
-import {baseLanguage} from './localeTitle'
+import {baseLanguage} from '../plugins/TranslationService/schemas/localeTitle'
 
 const isReadOnly = ({parent}: ConditionalPropertyCallbackContext) => {
   const {translationProcessing} = parent || {}
@@ -13,9 +12,6 @@ export default defineType({
   name: 'post',
   title: 'Post',
   type: 'document',
-  components: {
-    input: PostDocumentInput,
-  },
   icon: ComposeIcon,
   groups: [
     {
